requirements.txt
fastapi[all]
uvicorn
langchain
langchain-community
langchain-openai
azure-identity
azure-ai-ml
redis>=5.0.0
redisvl>=0.2.5
pydantic
python-dotenv
sqlalchemy
asyncpg
watchdog
python-multipart
Pillow
torch
torchaudio
transformers
llmlingua 
langchain-redis 
chromadb 
PyPDF2
fastapi-cors
httpx
pymongo
python-jose
passlib[bcrypt]
moz_sql_parser